requirements.txt
bcrypt==3.2.0
boto3==1.17.17
botocore==1.20.17
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
Deprecated==1.2.11
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
fastapi-users==5.1.0
h11==0.12.0
idna==3.1
jmespath==0.10.0
makefun==1.9.5
motor==2.3.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8
PyJWT==2.0.1
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.7.2
s3transfer==0.3.4
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.4
wrapt==1.12.1
